Updated Combat Arts to 3.0 and Custom Unit Menu Window to 2.4
The Combat Arts plugin was reworked, and all projects using an older version will need to tweak some of the custom data.
The Custom Unit Menu Window was updated to support the new version of the Combat Arts plugin.  Any project that was using this plugin must update to 2.4 when using the 3.0 update of the Combat Arts plugin.